egglog_concurrency

Module parallel_writer

A Utility Struct for Writing to a Vector in parallel without blocking reads.

Structs

ParallelVecWriter

A struct that wraps a vector and allows for parallel writes to it.

UnsafeReadAccess

A handle that can be used to read arbitrary locations in a vector wrapped by a ParallelVecWriter, even if they weren’t initialized when the ParallelVecWriter was created.

Functions

write_cell_slice

A variant of write_slice that applies to Cell<T>.