egglog_concurrency::parallel_writer

Function write_cell_slice

pub fn write_cell_slice<T: Copy>(
    this: &ParallelVecWriter<Cell<T>>,
    items: &[Cell<T>],
) -> usize

A variant of write_slice that applies to Cell<T>.

Cell<T> could just be copy but it is not for some reason0. All of the safety guarantees for slices of copy types apply to slices of cells just as well.

There seems to be a bug preventing this from being a member method of ParallelVecWriter.